Undo changes related to androauto
