+++ androguard/cli/__init__.py
@@ -1,9 +1,7 @@
 from androguard.cli.main import (androarsc_main,
-                                 androauto_main,
                                  androaxml_main,
                                  androcg_main,
                                  androgui_main,
-                                 AndroLog,
                                  androlyze_main,
                                  androsign_main,
                                  export_apps_to_format,
