+++ setup.py
@@ -71,7 +71,6 @@ setup(
     scripts=['androaxml.py',
              'androarsc.py',
              'androsign.py',
-             'androauto.py',
              'androdis.py',
              'androlyze.py',
              'androdd.py',
